+++ src/near_pass_detection/near_pass_prediction_params.h
@@ -3,9 +3,9 @@
 #define NEAR_PASS_PREDICTION_PARAMS_H
 
 const float MIN_SPEED_mps = 4;
-const float MIN_SPEED_MAGNITUDE = 25;
+const int MIN_SPEED_MAGNITUDE = 25;
 
 const float MIN_RANGE_m = 2.0;
-const float MIN_RANGE_MAGNITUDE = 300;
+const int MIN_RANGE_MAGNITUDE = 300;
 
 #endif
